moved api folder to utils
