--- packages/client/src/api/api.types.ts
@@ -1,6 +1,0 @@
-import type { SerializedError } from '@reduxjs/toolkit';
-import type { FetchBaseQueryError } from '@reduxjs/toolkit/dist/query';
-
-export type RTKQueryFetchError = FetchBaseQueryError;
-
-export type RTKQueryError = RTKQueryFetchError | SerializedError;
