+++ packages/client/src/types/index.ts
@@ -1,11 +1,9 @@
-import type {
-  RTKQueryError as TypeRTKQueryError,
-  RTKQueryFetchError as TypeRTKQueryFetchError,
-} from '@/api/api.types';
-
+import { SerializedError } from '@reduxjs/toolkit';
+import type { FetchBaseQueryError } from '@reduxjs/toolkit/dist/query';
 declare global {
-  export type RTKQueryFetchError = TypeRTKQueryFetchError;
-  export type RTKQueryError = TypeRTKQueryError;
+  export type RTKQueryFetchError = FetchBaseQueryError;
+
+  export type RTKQueryError = FetchBaseQueryError | SerializedError;
 }
 
 export {};
