+++ packages/client/src/utils/api/signup/error-messages/get-error-message.ts
@@ -1,4 +1,4 @@
-import { SignUpRequestErrorResponse } from '../signup-api.types';
+import { SignUpRequestErrorResponse } from '@/store/auth/auth.types';
 import { errorClassifiers } from './errors-classifiers';
 
 type Args = {
